+++ src/__tests__/core/web_api/helper.test.js
@@ -1,26 +1,27 @@
-import expect from 'expect.js';
 import { webAuthOverrides, normalizeError } from 'core/web_api/helper';
 
 describe('webAuthOverrides', () => {
   it('should return overrides if any field is compatible with WebAuth', function() {
-    expect(webAuthOverrides({ __tenant: 'tenant1', __token_issuer: 'issuer1' })).to.eql({
-      __tenant: 'tenant1',
-      __token_issuer: 'issuer1'
-    });
+    expect(webAuthOverrides({ __tenant: 'tenant1', __token_issuer: 'issuer1' })).toMatchSnapshot();
   });
 
   it('should omit overrides that are not compatible with WebAuth', function() {
     expect(
       webAuthOverrides({ __tenant: 'tenant1', __token_issuer: 'issuer1', backgroundColor: 'blue' })
-    ).to.eql({ __tenant: 'tenant1', __token_issuer: 'issuer1' });
+    ).toMatchSnapshot();
   });
 
   it('should return null if no fields are compatible with WebAuth', function() {
-    expect(webAuthOverrides({ backgroundColor: 'blue' })).to.not.be.ok();
+    expect(webAuthOverrides({ backgroundColor: 'blue' })).toBe(null);
   });
 });
+
 describe('normalizeError', () => {
-  it('should map access_denied error to invalid_user_password', () => {
+  it('does nothing when there is no error', () => {
+    const normalized = normalizeError(undefined);
+    expect(normalized).toBe(undefined);
+  });
+  it('should map access_denied error to invalid_user_password when error.error === access_denied', () => {
     const error = {
       error: 'access_denied',
       description: 'foobar'
@@ -28,9 +29,22 @@ describe('normalizeError', () => {
     const expectedError = {
       code: 'invalid_user_password',
       error: 'invalid_user_password',
-      description: error.description
+      description: 'foobar'
     };
     const actualError = normalizeError(error);
-    expect(actualError).to.be.eql(expectedError);
+    expect(actualError).toMatchSnapshot();
+  });
+  it('should map access_denied error to invalid_user_password when error.code === access_denied', () => {
+    const error = {
+      code: 'access_denied',
+      description: 'foobar'
+    };
+    const expectedError = {
+      code: 'invalid_user_password',
+      error: 'invalid_user_password',
+      description: 'foobar'
+    };
+    const actualError = normalizeError(error);
+    expect(actualError).toMatchSnapshot();
   });
 });