+++ src/__tests__/testUtils.js
@@ -53,3 +53,10 @@ export const setURL = url => {
     });
   });
 };
+
+export const expectMockToMatch = ({ mock }, numberOfCalls) => {
+  expect(mock.calls.length).toBe(numberOfCalls);
+  for (var i = 0; i < numberOfCalls; i++) {
+    expect(mock.calls[i]).toMatchSnapshot();
+  }
+};