+++ src/core/web_api/p2_api.js
@@ -1,5 +1,6 @@
 import auth0 from 'auth0-js';
 import CordovaAuth0Plugin from 'auth0-js/plugins/cordova';
+import superagent from 'superagent';
 import * as l from '../index';
 import { getEntity, read } from '../../store/index';
 import { normalizeError, loginCallback, normalizeAuthParams, webAuthOverrides } from './helper';
@@ -9,6 +10,8 @@ class Auth0APIClient {
     this.lockID = lockID;
     this.client = null;
     this.authOpt = null;
+    this.domain = domain;
+    this.isUniversalLogin = window.location.host === domain;
 
     const default_telemetry = {
       name: 'lock.js',
@@ -24,7 +27,7 @@ class Auth0APIClient {
       responseMode: opts.responseMode,
       responseType: opts.responseType,
       leeway: opts.leeway || 1,
-      plugins: [new CordovaAuth0Plugin()],
+      plugins: opts.plugins || [new CordovaAuth0Plugin()],
       overrides: webAuthOverrides(opts.overrides),
       _sendTelemetry: opts._sendTelemetry === false ? false : true,
       _telemetryInfo: opts._telemetryInfo || default_telemetry
@@ -36,6 +39,9 @@ class Auth0APIClient {
       nonce: opts.nonce,
       state: opts.state
     };
+    if (this.isUniversalLogin && opts.sso !== undefined) {
+      this.authOpt.sso = opts.sso;
+    }
   }
 
   logIn(options, authParams, cb) {
@@ -75,8 +81,13 @@ class Auth0APIClient {
     this.client.changePassword(options, cb);
   }
 
-  startPasswordless(options, cb) {
-    this.client.startPasswordless(options, err => cb(normalizeError(err)));
+  passwordlessStart(options, cb) {
+    this.client.passwordlessStart(options, err => cb(normalizeError(err)));
+  }
+
+  passwordlessVerify(options, cb) {
+    const verifyOptions = { ...options, popup: this.authOpt.popup };
+    this.client.passwordlessLogin(verifyOptions, (err, result) => cb(normalizeError(err), result));
   }
 
   parseHash(hash = '', cb) {
@@ -98,12 +109,21 @@ class Auth0APIClient {
     this.getUserInfo(token, callback);
   }
 
-  getSSOData(...args) {
-    return this.client.client.getSSOData(...args);
+  getSSOData(cb) {
+    if (this.isUniversalLogin) {
+      superagent
+        .get(`https://${this.domain}/user/ssodata`)
+        .withCredentials()
+        .end((err, res) => {
+          cb(err, res.body);
+        });
+    } else {
+      return this.client.client.getSSOData(cb);
+    }
   }
 
   getUserCountry(cb) {
-    return this.client.getUserCountry(cb);
+    return this.client.client.getUserCountry(cb);
   }
 
   checkSession(options, cb) {